+++ src/ppp_ar.c
@@ -15,6 +15,6 @@
 
 /* ambiguity resolution in ppp -----------------------------------------------*/
 extern int ppp_ar(rtk_t *rtk, const obsd_t *obs, int n, int *exc, const nav_t *nav,
-                  const double *azel, double *x, double *P) {
+                  const long double *azel, long double *x, long double *P) {
   return 0;
 }
